Add first failing test to suite. Also fix typing error in checkout signature

+++ src/main/scala/Checkout.scala
@@ -1,4 +1,4 @@
 
 object Checkout {
-  def checkout(items: List[Fruit]): Int = ??? //Unimplemented - awaiting TDD
+  def checkout(items: List[Fruit]): Double = ??? //Unimplemented - awaiting TDD
 }
